moves jwt auth_func from tools.utils into libs.auth
